refactor: split test files by fixture
