+++ packages/integrations/markdoc/test/render-components.test.js
@@ -3,82 +3,58 @@ import { describe, it, before, after } from 'node:test';
 import { parseHTML } from 'linkedom';
 import { loadFixture } from '../../../astro/test/test-utils.js';
 
-async function getFixture(name) {
-	return await loadFixture({
-		root: new URL(`./fixtures/${name}/`, import.meta.url),
-	});
-}
+const root = new URL('./fixtures/render-with-components/', import.meta.url);
 
 describe('Markdoc - render components', () => {
+	let fixture;
+
+	before(async () => {
+		fixture = await loadFixture({
+			root,
+		});
+	});
+
 	describe('dev', () => {
-		let componentsFixture;
-		let componentsServer;
+		let devServer;
 
 		before(async () => {
-			componentsFixture = await getFixture('render-with-components');
-			componentsServer = await componentsFixture.startDevServer();
+			devServer = await fixture.startDevServer();
 		});
 
 		after(async () => {
-			await componentsServer.stop();
+			await devServer.stop();
 		});
 
 		it('renders content - with components', async () => {
-			const res = await componentsFixture.fetch('/');
+			const res = await fixture.fetch('/');
 			const html = await res.text();
 
 			renderComponentsChecks(html);
 		});
 
 		it('renders content - with components inside partials', async () => {
-			const res = await componentsFixture.fetch('/');
-			const html = await res.text();
-
-			renderComponentsInsidePartialsChecks(html);
-		});
-
-		it('renders content - with indented components', async () => {
-			const fixture = await getFixture('render-with-indented-components');
-			const server = await fixture.startDevServer();
-
 			const res = await fixture.fetch('/');
 			const html = await res.text();
 
-			renderIndentedComponentsChecks(html);
-
-			await server.stop();
+			renderComponentsInsidePartialsChecks(html);
 		});
 	});
 
 	describe('build', () => {
-		let componentsFixture;
-
 		before(async () => {
-			componentsFixture = await getFixture('render-with-components');
-		});
-
-		before(async () => {
-			await componentsFixture.build();
+			await fixture.build();
 		});
 
 		it('renders content - with components', async () => {
-			const html = await componentsFixture.readFile('/index.html');
+			const html = await fixture.readFile('/index.html');
 
 			renderComponentsChecks(html);
 		});
 
 		it('renders content - with components inside partials', async () => {
-			const html = await componentsFixture.readFile('/index.html');
-
-			renderComponentsInsidePartialsChecks(html);
-		});
-
-		it('renders content - with indented components', async () => {
-			const fixture = await getFixture('render-with-indented-components');
-			await fixture.build();
 			const html = await fixture.readFile('/index.html');
 
-			renderIndentedComponentsChecks(html);
+			renderComponentsInsidePartialsChecks(html);
 		});
 	});
 });
@@ -110,24 +86,4 @@ function renderComponentsInsidePartialsChecks(html) {
 	// renders DeeplyNested.astro
 	const deeplyNested = document.querySelector('#deeply-nested');
 	assert.equal(deeplyNested.textContent, 'Deeply nested partial');
-}
-
-/** @param {string} html */
-function renderIndentedComponentsChecks(html) {
-	const { document } = parseHTML(html);
-	const h2 = document.querySelector('h2');
-	assert.equal(h2.textContent, 'Post with indented components');
-
-	// Renders custom shortcode components
-	const marquees = document.querySelectorAll('marquee');
-	assert.equal(marquees.length, 2);
-
-	// Renders h3
-	const h3 = document.querySelector('h3');
-	assert.equal(h3.textContent, 'I am an h3!');
-
-	// Renders Astro Code component
-	const pre = document.querySelector('pre');
-	assert.notEqual(pre, null);
-	assert.equal(pre.className, 'astro-code github-dark');
 }
